Implement output buffering as Middleware
Create \Slim\Middleware\OutputBuffer which should be added as the
innermost middleware.

This middleware will ensure that echo calls in a route callable are
captured in an output buffer and appended to the response object. If a
response object is returned from the route callable, then this new
object is used for appending to and is returned. This allows you to echo
from a route callable, but also set the status code. We also allow for
returning a string fro the route callable and appending that to the
response.